requirements.txt
Django>=5.2.8
djangorestframework>=3.16.1
boto3>=1.34.0
botocore>=1.34.0
psycopg>=3.2.13
python-dotenv>=1.0.0

